+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # ESPresenseController Changelog
 
+## 230103
+
+* (Breaking Change) All ids are now normalized and lowercase'd.
+* Mark as dead and purge entities not reporting after 5 days
+
 ## 230102
 
 * Initial release version.